requirements.txt
tensorflow
numpy
pandas
google-api-python-client
python-dotenv
flask
flask_cors
Flask-Session
openpyxl